HAPP-1060: Refactor ForecastStatus
