+++ app/src/main/java/ca/bc/gov/bchealth/ui/recommendations/RecommendationsViewModel.kt
@@ -2,6 +2,7 @@ package ca.bc.gov.bchealth.ui.recommendations
 
 import androidx.lifecycle.ViewModel
 import ca.bc.gov.bchealth.model.mapper.toUiModel
+import ca.bc.gov.common.model.immunization.ForecastStatus
 import ca.bc.gov.repository.immunization.ImmunizationRecommendationRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.map
@@ -19,7 +20,7 @@ class RecommendationsViewModel @Inject constructor(
 
 data class RecommendationDetailItem(
     val title: String,
-    val status: String?,
+    val status: ForecastStatus?,
     val date: String,
     var fullContent: Boolean = false,
 )
