+++ common/src/main/java/ca/bc/gov/common/model/immunization/ImmunizationForecastDto.kt
@@ -16,21 +16,3 @@ data class ImmunizationForecastDto(
     val dueDate: Instant
 )
 
-sealed class ForecastStatus(open val text: String) {
-    object UpToDate : ForecastStatus("Up to date") // up to date in series
-    object Eligible : ForecastStatus("Eligible") // eligible date has passed
-    object Due : ForecastStatus("Due") // within 2 weeks of due date
-    object Overdue : ForecastStatus("Overdue") // Overdue: past due date
-    data class Other(override val text: String) : ForecastStatus(text)
-
-    companion object {
-        fun getByText(text: String?): ForecastStatus? = when (text?.lowercase()) {
-            UpToDate.text.lowercase() -> UpToDate
-            Eligible.text.lowercase() -> Eligible
-            Due.text.lowercase() -> Due
-            Overdue.text.lowercase() -> Overdue
-            null -> null
-            else -> Other(text)
-        }
-    }
-}
