+++ data/src/main/java/ca/bc/gov/data/model/mapper/DtoToEntityMapper.kt
@@ -239,6 +239,6 @@ fun ImmunizationRecommendationsDto.toEntity() = ImmunizationRecommendationEntity
     recommendationSetId = this.recommendationSetId,
     patientId = this.patientId,
     immunizationName = this.immunizationName,
-    status = this.status,
+    status = this.status?.text,
     diseaseDueDate = this.diseaseDueDate,
 )
